Fixed frequencies in makeW

+++ src/ide_helpers.h
@@ -17,9 +17,9 @@ void makeB(arma::mat & B, const arma::mat & mu, const arma::cube & Sigma,
            const arma::mat & locs, const arma::mat & w, const int J, const int L);
 
 arma::mat makeF(const arma::mat & locs, const arma::mat & w,
-                  const int J, const int L);
+                const int J, const double L);
 
-arma::mat makeW(const int J, const int L);
+arma::mat makeW(const int J, const double L);
 
 arma::mat proposeMu(arma::mat mu, arma::mat Sigma);
